Move DeprecatedEnum to deprecation_utils
