+++ src/datasets/utils/deprecation_utils.py
@@ -1,3 +1,4 @@
+import enum
 import warnings
 from functools import wraps
 from typing import Callable, Optional
@@ -41,3 +42,52 @@ def deprecated(help_message: Optional[str] = None):
         return wrapper
 
     return decorator
+
+
+class OnAccess(enum.EnumMeta):
+    """
+    Enum metaclass that calls a user-specified function whenever a member is accessed.
+    """
+
+    def __getattribute__(cls, name):
+        obj = super().__getattribute__(name)
+        if isinstance(obj, enum.Enum) and obj._on_access:
+            obj._on_access()
+        return obj
+
+    def __getitem__(cls, name):
+        member = super().__getitem__(name)
+        if member._on_access:
+            member._on_access()
+        return member
+
+    def __call__(cls, value, names=None, *, module=None, qualname=None, type=None, start=1):
+        obj = super().__call__(value, names, module=module, qualname=qualname, type=type, start=start)
+        if isinstance(obj, enum.Enum) and obj._on_access:
+            obj._on_access()
+        return obj
+
+
+class DeprecatedEnum(enum.Enum, metaclass=OnAccess):
+    """
+    Enum class that calls `deprecate` method whenever a member is accessed.
+    """
+
+    def __new__(cls, value):
+        member = object.__new__(cls)
+        member._value_ = value
+        member._on_access = member.deprecate
+        return member
+
+    @property
+    def help_message(self):
+        return ""
+
+    def deprecate(self):
+        help_message = f" {self.help_message}" if self.help_message else ""
+        warnings.warn(
+            f"'{self.__objclass__.__name__}' is deprecated and will be removed in the next major version of datasets."
+            + help_message,
+            FutureWarning,
+            stacklevel=3,
+        )
