+++ src/datasets/utils/py_utils.py
@@ -18,7 +18,6 @@
 """
 
 import contextlib
-import enum
 import functools
 import itertools
 import os
@@ -26,7 +25,6 @@ import pickle
 import re
 import sys
 import types
-import warnings
 from io import BytesIO as StringIO
 from multiprocessing import Pool, RLock
 from shutil import disk_usage
@@ -638,52 +636,3 @@ try:
 
 except ImportError:
     pass
-
-
-class OnAccess(enum.EnumMeta):
-    """
-    Enum metaclass that calls a user-specified function whenever a member is accessed.
-    """
-
-    def __getattribute__(cls, name):
-        obj = super().__getattribute__(name)
-        if isinstance(obj, enum.Enum) and obj._on_access:
-            obj._on_access()
-        return obj
-
-    def __getitem__(cls, name):
-        member = super().__getitem__(name)
-        if member._on_access:
-            member._on_access()
-        return member
-
-    def __call__(cls, value, names=None, *, module=None, qualname=None, type=None, start=1):
-        obj = super().__call__(value, names, module=module, qualname=qualname, type=type, start=start)
-        if isinstance(obj, enum.Enum) and obj._on_access:
-            obj._on_access()
-        return obj
-
-
-class DeprecatedEnum(enum.Enum, metaclass=OnAccess):
-    """
-    Enum class that calls `deprecate` method whenever a member is accessed.
-    """
-
-    def __new__(cls, value):
-        member = object.__new__(cls)
-        member._value_ = value
-        member._on_access = member.deprecate
-        return member
-
-    @property
-    def help_message(self):
-        return ""
-
-    def deprecate(self):
-        help_message = f" {self.help_message}" if self.help_message else ""
-        warnings.warn(
-            f"'{self.__objclass__.__name__}' is deprecated and will be removed in the next major version of datasets."
-            + help_message,
-            DeprecationWarning,
-            stacklevel=3,
-        )
